All functionality implemented!  Yes, it compiles!  May have some bugs yet, but it's all there!

+++ h/uartDrv.h
@@ -12,7 +12,7 @@
 
 //function prototypes
 extern void uartInit (void);
-extern void transmitResults(uint8_t sensor, double *phaseAngle, float *amplitude, bool bridgeADCClipFlag, bool coilADCClipFlag, bool bridgeDigitalClipFlag);
+extern void transmitResults(uint8_t sensor, float *phaseAngle, float *amplitude, bool bridgeADCClipFlag, bool coilADCClipFlag, bool bridgeDigitalClipFlag);
 extern void transmitError(uint8_t errorCode);
 
 //global variables
